requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.2
PyYAML==6.0
mergedeep==1.3.4
requests
rich==13.7.1